cyberDefense/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0
matplotlib>=3.4.0
seaborn>=0.11.0
scapy>=2.4.5
requests>=2.27.0
cryptography>=36.0.0
psutil>=5.9.0
flask>=2.0.0
python-dotenv>=0.19.0
joblib>=1.1.0
